[INFRA] main 브랜치로 merge (#26)
* feat: 카카오 일정 조회 구현 중

* fix: import 임시 주석 처리

* feat: 일정 등록시 카카오 API 조회

* feat: 카카오 장소 조회 기능 추가

* feat: 네이버 블로그 조회 기능 추가

* feat: 태그번호로 태그명 조회 기능

* feat: ai 추천 로직 진행중

* feat: ai 일정 추천 기능 진행중

* feat: ai 일정 추천 결과 리턴하도록 작업

* chore: 필요 없는 로그 주석처리

* fix: tag 조회 쿼리 테이블 명 대문자로 수정(소문자 사용 시 에러)

* chore: AIClient 필요없는 주석 제거

* feat: 주소를 x,y 좌표로 변환하는 api 추가

* feat: ApiResponse에 success/error 정적팩토리 메서드 추가

* feat: 주소 목록 검색 api 추가

* fix: naverBlog 패키지 이름 수정

* fix: tagList를 int 배열이 아니라 map 형태로 보내주도록 수정

* fix: 일정 생성 기능 수정

* feat: 생성된 일정을 DB에 저장하는 API 작업

* update: 일정 상세 조회 API의 태그 조회 쿼리 수정

* feat: 내 일정 목록 조회 API 작업

* update: Plan, Schedule 엔티티에 등록일, 수정일, 삭제여부 컬럼 추가

* update: Plan, Schedule 저장시 삭제여부 함께 저장, 일정 저장 API 메서드명 수정

* feat: 전체 일정 삭제 API 작업

* feat: 일정 수정, 부분삭제 API 작업

* feat: 카카오 장소 키워드 검색 API 추가

* fix: 일정 수정 시 soft-delete 하도록 변경

* feat: 사용자가 직접 추가한 일정, 카카오 API로 검색한 일정 생성할 수 있도록 수정

* fix: 빌드 에러 해결

* Fix/member api (#21)

* FIX
1. 일반 회원가입 관련 API return 값을 swagger에 상세히 작성
2. 회원가입 정보 저장 API 에 아이디 중복 체크 추가

* FIX
1. 약관 정보 등록 API 수정

* FIX
1. 일반 로그인 관련 API 수정

* FIX
1. 회원 정보 수정 API 수정

* FIX
1. oauth, login 관련 코드 수정

* FIX
1. oauth, login 관련 코드 수정

* [FIX] 인기 태그 조회 API 수정 (#22)

1. 인기 태그 조회 API 수정
- 태그가 최대 10개만 조회되도록 수정

* [FIX] 회원 정보 관련 API 수정 (#23)

- 프로필 이미지 저장되지 않도록 수정
- 프로필 이미지 조회 불가능하도록 수정
- 회원 정보 수정 시 닉네임, 성별, 생년월일만 수정되도록 수정

* [FIX] 회원 탈퇴 코드 수정 (#24)

* [FIX] 회원 탈퇴 코드 수정
- access token이 아닌 refresh token 값으로 회원 탈퇴 처리

* [FIX] 회원 탈퇴 코드 수정
- token 이용해서 db 조회 전에 검증

* feat: 일정 생성 시 랜덤 카테고리를 선택할 수 있도록 수정 (#25)

---------

+++ src/main/java/com/barogagi/member/login/controller/AuthController.java
@@ -4,13 +4,13 @@ import com.barogagi.member.login.dto.*;
 import com.barogagi.member.login.exception.InvalidRefreshTokenException;
 import com.barogagi.member.login.service.AccountService;
 import com.barogagi.member.login.service.AuthService;
+import com.barogagi.response.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
@@ -26,7 +26,8 @@ public class AuthController {
     private final AuthService authService;
     private final AccountService accountService;
 
-    public AuthController(AuthService authService, AccountService accountService) {
+    public AuthController(AuthService authService,
+                          AccountService accountService) {
         this.authService = authService;
         this.accountService = accountService;
     }
@@ -34,9 +35,8 @@ public class AuthController {
     @Operation(summary = "토큰 재발급", description = "Access 토큰 만료 시, Refresh 토큰으로 Access 토큰 재발급")
     @PostMapping("/refresh")
     public ResponseEntity<Map<String, Object>> refresh(
-            @RequestHeader(value = "X-Refresh-Token", required = false) String refreshHeader,
-            @RequestBody(required = false) Map<String, String> body,
-            HttpServletResponse response
+            @RequestHeader(value = "Refresh-Token", required = false) String refreshHeader,
+            @RequestBody(required = false) Map<String, String> body
     ) {
 
         logger.info("CALL /auth/refresh");
@@ -73,13 +73,13 @@ public class AuthController {
     /**
      * 현재 기기 로그아웃: 전달된 refreshToken이 속한 (membershipNo, deviceId)의 VALID 토큰들을 REVOKE
      * 입력 경로:
-     *  - 헤더: X-Refresh-Token: <refresh>
+     *  - 헤더: Refresh-Token: <refresh>
      *  - 바디: { "refreshToken": "<refresh>" }
      */
     @Operation(summary = "현재 기기 로그아웃", description = "현재 기기 로그아웃: 전달된 refreshToken이 속한 (membershipNo, deviceId)의 VALID 토큰들을 REVOKE")
     @PostMapping("/logout")
     public ResponseEntity<Map<String, Object>> logout(
-            @RequestHeader(value = "X-Refresh-Token", required = false) String refreshHeader,
+            @RequestHeader(value = "Refresh-Token", required = false) String refreshHeader,
             @RequestBody(required = false) Map<String, String> body
     ) {
         String refresh = refreshHeader;
@@ -89,14 +89,69 @@ public class AuthController {
         if (refresh != null && !refresh.isBlank()) {
             authService.logout(refresh); // DB REVOKE
         }
-        return ResponseEntity.ok(Map.of("result", "logged_out"));
+        return ResponseEntity.ok(Map.of("resultCode", "200", "message", "로그아웃 되었습니다."));
     }
 
-    @DeleteMapping
-    public ResponseEntity<?> deleteMe(Authentication auth) {
-        String membershipNo = (String) auth.getPrincipal(); // JwtAuthFilter에서 세팅됨
-        accountService.deleteMyAccount(membershipNo);
-        return ResponseEntity.noContent().build(); // 204
+    @Operation(summary = "회원 탈퇴", description = "회원 탈퇴 API",
+            responses =  {
+                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "100", description = "refresh token이 존재하지 않습니다."),
+                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "회원 탈퇴되었습니다."),
+                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "300", description = "회원 탈퇴 실패하였습니다."),
+                    @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "400", description = "오류가 발생하였습니다.")
+            })
+    @PostMapping("/member/delete")
+    public ApiResponse deleteMe(@RequestHeader(value = "Refresh-Token", required = false) String refreshHeader,
+                                @RequestBody(required = false) Map<String, String> body) {
+
+        logger.info("CALL /auth/member/delete");
+
+        ApiResponse apiResponse = new ApiResponse();
+        String resultCode = "";
+        String message = "";
+
+        try {
+
+            String refreshToken = Optional.ofNullable(refreshHeader)
+                    .or(() -> Optional.ofNullable(body == null ? null : body.get("refreshToken")))
+                    .orElse(null);
+
+            if (refreshToken == null || refreshToken.isBlank()) {
+                throw new InvalidRefreshTokenException("100", "refresh token이 존재하지 않습니다.");
+            }
+
+            // refresh token을 이용해서 membershipNo 구하기
+
+            Map<String, String> resultMap = authService.selectUserInfoByToken(refreshToken);
+            if(!resultMap.get("resultCode").equals("200")) {
+                throw new InvalidRefreshTokenException(resultMap.get("resultCode"), resultMap.get("message"));
+            }
+
+            String membershipNo = resultMap.get("membershipNo");
+
+            int deleteResult = accountService.deleteMyAccount(membershipNo);
+            if(deleteResult > 0) {
+                resultCode = "200";
+                message = "회원 탈퇴되었습니다.";
+            } else {
+                resultCode = "300";
+                message = "회원 탈퇴 실패하였습니다.";
+            }
+
+        } catch (InvalidRefreshTokenException ex) {
+            resultCode = ex.getCode();
+            message = ex.getMessage();
+
+        } catch (Exception e) {
+            logger.error("error", e);
+            resultCode = "400";
+            message = "오류가 발생하였습니다.";
+
+        } finally {
+            apiResponse.setResultCode(resultCode);
+            apiResponse.setMessage(message);
+        }
+
+        return apiResponse;
     }
 }
